+++ test/integration/flows/child_benefit_tax_calculator_test.rb
@@ -2,7 +2,7 @@
 require_relative '../../test_helper'
 require_relative 'flow_test_helper'
 
-class ChildBenefitTaxCalculatorTest < ActiveSupport::TestCase
+class ChildBenefitTaxCalculatorV2Test < ActiveSupport::TestCase
   include FlowTestHelper
 
   setup do
@@ -18,11 +18,11 @@ class ChildBenefitTaxCalculatorTest < ActiveSupport::TestCase
     setup do
       add_response :just_how_much
     end
-    
+
     should "ask which tax year you want an estimate for" do
       assert_current_node :which_tax_year?
     end
-    
+
     context "enter 2013-14" do
       setup do
         add_response "2013-14"
@@ -203,7 +203,7 @@ class ChildBenefitTaxCalculatorTest < ActiveSupport::TestCase
               end
 
               should "calculate adjusted net income" do
-                assert_state_variable :adjusted_net_income, 19410 
+                assert_state_variable :adjusted_net_income, 19410
               end
 
               should "not require to pay tax when adjusted net income less than £50,000" do
@@ -324,8 +324,8 @@ class ChildBenefitTaxCalculatorTest < ActiveSupport::TestCase
                     assert_current_node_is_error
                   end
 
-                  should "reject > 3" do
-                    add_response "4"
+                  should "reject > 9" do
+                    add_response "10"
                     assert_current_node :how_many_children_to_start_claiming?
                     assert_current_node_is_error
                   end
@@ -340,10 +340,9 @@ class ChildBenefitTaxCalculatorTest < ActiveSupport::TestCase
                     end
 
                     should "reject a date outside of the current tax year" do
-                      add_response "2013-11-05" do
-                        assert_current_node :when_do_you_expect_to_stop_claiming_for_the_1st_child?
-                        assert_current_node_is_error
-                      end
+                      add_response "2013-11-05"
+                      assert_current_node :when_will_the_1st_child_enter_the_household?
+                      assert_current_node_is_error
                     end
 
                     context "valid first child date" do
@@ -351,100 +350,126 @@ class ChildBenefitTaxCalculatorTest < ActiveSupport::TestCase
                         add_response "2012-12-22"
                       end
 
-                      should "ask the starting date of the second child, given the first child start date" do
-                        assert_current_node :when_will_the_2nd_child_enter_the_household?
+                      should "ask if the first child will be leaving the household this year" do
+                        assert_current_node :will_the_1st_child_leave_the_household_this_year?
                       end
 
-                      context "valid second child date" do
+                      context "not leaving this year" do
                         setup do
-                          add_response "2013-02-14"
+                          add_response 'no'
                         end
 
-                        should "ask how many children you expect to stop claiming for" do
-                          assert_current_node :how_many_children_to_stop_claiming?
+                        should "ask the starting date of the second child, given the first child start date" do
+                          assert_current_node :when_will_the_2nd_child_enter_the_household?
                         end
 
-                        should "reject > number of children" do
-                          add_response "6"
-                          assert_current_node :how_many_children_to_stop_claiming?
-                          assert_current_node_is_error
-                        end
-
-                        context "no children stopping" do
+                        context "valid second child date" do
                           setup do
-                            add_response "0"
+                            add_response "2013-02-14"
                           end
 
-                          should "tell you your estimated tax charge" do
-                            SmartAnswer::Calculators::ChildBenefitTaxCalculator.
-                              expects(:new).with(
-                                :start_of_tax_year => Date.new(2012, 4, 6),
-                                :end_of_tax_year => Date.new(2013, 4, 5),
-                                :children_claiming => 1,
-                                :claim_periods => [Date.new(2012,12,22)..Date.new(2013,4,5), Date.new(2013,2,14)..Date.new(2013,4,5), Date.new(2012,4,6)..Date.new(2013,4,5)],
-                                :income => 50610.0
-                              ).returns(@stubbed_calculator)
-                            @stubbed_calculator.expects(:formatted_benefit_tax).returns("formatted benefit tax")
-                            @stubbed_calculator.expects(:formatted_benefit_taxable_amount).returns("formatted benefit taxable amount")
-                            @stubbed_calculator.expects(:formatted_benefit_claimed_amount).returns("formatted benefit claimed amount")
-                            @stubbed_calculator.expects(:benefit_taxable_weeks).returns("benefit taxable weeks")
-                            @stubbed_calculator.expects(:percent_tax_charge).returns("percent tax charge")
-
-                            assert_current_node :estimated_tax_charge
-                            assert_state_variable :benefit_taxable_amount, "formatted benefit taxable amount"
-                            assert_state_variable :benefit_claimed_amount, "formatted benefit claimed amount"
-                            assert_state_variable :percentage_tax_charge, "percent tax charge"
-                            assert_state_variable :benefit_taxable_weeks, "benefit taxable weeks"
-                            assert_state_variable :benefit_tax, "formatted benefit tax"
-                          end
-                        end # context - no children stopping
-
-                        context "1 child stopping" do
-                          setup do
-                            add_response "1"
+                          should "ask if the second child will be leaving the household this year" do
+                            assert_current_node :will_the_2nd_child_leave_the_household_this_year?
                           end
 
-                          should "ask the stopping date of the first child" do
-                            assert_current_node :when_do_you_expect_to_stop_claiming_for_the_1st_child?
-                          end
+                          context "not leaving this year" do
+                            setup do
+                              add_response 'no'
+                            end
 
-                          should "reject a date outside of the current tax year" do
-                            add_response "2014-01-10" do
-                              assert_current_node :when_do_you_expect_to_stop_claiming_for_the_1st_child?
+                            should "ask how many children you expect to stop claiming for" do
+                              assert_current_node :how_many_children_to_stop_claiming?
+                            end
+
+                            should "reject > number of children" do
+                              add_response "6"
+                              assert_current_node :how_many_children_to_stop_claiming?
                               assert_current_node_is_error
                             end
-                          end
 
-                          context "valid first child date" do
-                            setup do
-                              add_response "2013-01-15"
+                            should "reject non-numeric values" do
+                              add_response "foobar"
+                              assert_current_node :how_many_children_to_stop_claiming?
+                              assert_current_node_is_error
                             end
 
-                            should "tell you your estimated tax charge" do
-                              SmartAnswer::Calculators::ChildBenefitTaxCalculator.
-                                expects(:new).with(
-                                  :start_of_tax_year => Date.new(2012, 4, 6),
-                                  :end_of_tax_year => Date.new(2013, 4, 5),
-                                  :children_claiming => 1,
-                                  :claim_periods => [Date.new(2012,12,22)..Date.new(2013,4,5), Date.new(2013,2,14)..Date.new(2013,4,5), Date.new(2012,4,6)..Date.new(2013,1,15)],
-                                  :income => 50610.0
-                                ).returns(@stubbed_calculator)
-                              @stubbed_calculator.expects(:formatted_benefit_tax).returns("formatted benefit tax")
-                              @stubbed_calculator.expects(:formatted_benefit_taxable_amount).returns("formatted benefit taxable amount")
-                              @stubbed_calculator.expects(:formatted_benefit_claimed_amount).returns("formatted benefit claimed amount")
-                              @stubbed_calculator.expects(:benefit_taxable_weeks).returns("benefit taxable weeks")
-                              @stubbed_calculator.expects(:percent_tax_charge).returns("percent tax charge")
+                            context "no children stopping" do
+                              setup do
+                                add_response "0"
+                              end
 
-                              assert_current_node :estimated_tax_charge
-                              assert_state_variable :benefit_taxable_amount, "formatted benefit taxable amount"
-                              assert_state_variable :benefit_claimed_amount, "formatted benefit claimed amount"
-                              assert_state_variable :percentage_tax_charge, "percent tax charge"
-                              assert_state_variable :benefit_taxable_weeks, "benefit taxable weeks"
-                              assert_state_variable :benefit_tax, "formatted benefit tax"
-                            end
-                          end # context - valid first child date
-                        end # context - 1 child stopping
-                      end # context - valid second child date
+                              should "tell you your estimated tax charge" do
+                                SmartAnswer::Calculators::ChildBenefitTaxCalculator.
+                                  expects(:new).with(
+                                    :start_of_tax_year => Date.new(2012, 4, 6),
+                                    :end_of_tax_year => Date.new(2013, 4, 5),
+                                    :children_claiming => 1,
+                                    :claim_periods => [Date.new(2012,12,22)..Date.new(2013,4,5), Date.new(2013,2,14)..Date.new(2013,4,5), Date.new(2012,4,6)..Date.new(2013,4,5)],
+                                    :income => 50610.0
+                                  ).returns(@stubbed_calculator)
+                                @stubbed_calculator.expects(:formatted_benefit_tax).returns("formatted benefit tax")
+                                @stubbed_calculator.expects(:formatted_benefit_taxable_amount).returns("formatted benefit taxable amount")
+                                @stubbed_calculator.expects(:formatted_benefit_claimed_amount).returns("formatted benefit claimed amount")
+                                @stubbed_calculator.expects(:benefit_taxable_weeks).returns("benefit taxable weeks")
+                                @stubbed_calculator.expects(:percent_tax_charge).returns("percent tax charge")
+
+                                assert_current_node :estimated_tax_charge
+                                assert_state_variable :benefit_taxable_amount, "formatted benefit taxable amount"
+                                assert_state_variable :benefit_claimed_amount, "formatted benefit claimed amount"
+                                assert_state_variable :percentage_tax_charge, "percent tax charge"
+                                assert_state_variable :benefit_taxable_weeks, "benefit taxable weeks"
+                                assert_state_variable :benefit_tax, "formatted benefit tax"
+                              end
+                            end # context - no children stopping
+
+                            context "1 child stopping" do
+                              setup do
+                                add_response "1"
+                              end
+
+                              should "ask the stopping date of the first child" do
+                                assert_current_node :when_do_you_expect_to_stop_claiming_for_the_1st_child?
+                              end
+
+                              should "reject a date outside of the current tax year" do
+                                add_response "2014-01-10" do
+                                  assert_current_node :when_do_you_expect_to_stop_claiming_for_the_1st_child?
+                                  assert_current_node_is_error
+                                end
+                              end
+
+                              context "valid first child date" do
+                                setup do
+                                  add_response "2013-01-15"
+                                end
+
+                                should "tell you your estimated tax charge" do
+                                  SmartAnswer::Calculators::ChildBenefitTaxCalculator.
+                                    expects(:new).with(
+                                      :start_of_tax_year => Date.new(2012, 4, 6),
+                                      :end_of_tax_year => Date.new(2013, 4, 5),
+                                      :children_claiming => 1,
+                                      :claim_periods => [Date.new(2012,12,22)..Date.new(2013,4,5), Date.new(2013,2,14)..Date.new(2013,4,5), Date.new(2012,4,6)..Date.new(2013,1,15)],
+                                      :income => 50610.0
+                                    ).returns(@stubbed_calculator)
+                                  @stubbed_calculator.expects(:formatted_benefit_tax).returns("formatted benefit tax")
+                                  @stubbed_calculator.expects(:formatted_benefit_taxable_amount).returns("formatted benefit taxable amount")
+                                  @stubbed_calculator.expects(:formatted_benefit_claimed_amount).returns("formatted benefit claimed amount")
+                                  @stubbed_calculator.expects(:benefit_taxable_weeks).returns("benefit taxable weeks")
+                                  @stubbed_calculator.expects(:percent_tax_charge).returns("percent tax charge")
+
+                                  assert_current_node :estimated_tax_charge
+                                  assert_state_variable :benefit_taxable_amount, "formatted benefit taxable amount"
+                                  assert_state_variable :benefit_claimed_amount, "formatted benefit claimed amount"
+                                  assert_state_variable :percentage_tax_charge, "percent tax charge"
+                                  assert_state_variable :benefit_taxable_weeks, "benefit taxable weeks"
+                                  assert_state_variable :benefit_tax, "formatted benefit tax"
+                                end
+                              end # context - valid first child date
+                            end # context - 1 child stopping
+                          end # context - second child not leaving this year
+                        end # context - valid second child date
+                      end # context - first child not leaving this year
                     end # context - valid first child date
                   end # context - 2 children starting
                 end # context - starting or stopping this tax year
@@ -453,8 +478,103 @@ class ChildBenefitTaxCalculatorTest < ActiveSupport::TestCase
           end # context - valid net savings interest
         end # context - not paying into pension this year
       end # context - income rounded to £100 greater than £50,000
-    end # context - 2012-13 tax year 
+    end # context - 2012-13 tax year
   end
+
+  context "for foster carers" do
+    setup do
+      add_response :just_how_much
+      add_response "2013-14"
+    end
+
+    should "ask how many children" do
+      assert_current_node :how_many_children_claiming_for?
+    end
+
+    context "for 0 children" do
+      setup do
+        add_response "0"
+      end
+
+      should "ask if you expect to start or stop claiming during this tax year" do
+        assert_current_node :do_you_expect_to_start_or_stop_claiming?
+      end
+
+      context "starting or stopping this tax year" do
+        setup do
+          add_response :yes
+        end
+
+        should "ask how many children to start claiming" do
+          assert_current_node :how_many_children_to_start_claiming?
+        end
+
+        context "for 2 children starting and stopping this year" do
+          setup do
+            add_response "2"
+          end
+
+          should "ask when the first child is starting" do
+            assert_current_node :when_will_the_1st_child_enter_the_household?
+          end
+
+          context "entering on June 1st, 2013" do
+            setup do
+              add_response "2013-06-01"
+            end
+
+            should "ask if the first child will be leaving this year" do
+              assert_current_node :will_the_1st_child_leave_the_household_this_year?
+            end
+
+            context "leaving on May 1st, 2013" do
+              should "raise an error" do
+                add_response "2013-05-01"
+                assert_current_node :will_the_1st_child_leave_the_household_this_year?
+                assert_current_node_is_error
+              end
+            end
+
+            context "leaving on June 30th, 2013" do
+              setup do
+                add_response "2013-06-30"
+              end
+
+              should "ask when the second child is starting" do
+                assert_current_node :when_will_the_2nd_child_enter_the_household?
+              end
+
+              context "entering on September 1st, 2013" do
+                setup do
+                  add_response "2013-09-01"
+                end
+
+                should "ask if the second child will be leaving this year" do
+                  assert_current_node :will_the_2nd_child_leave_the_household_this_year?
+                end
+
+                context "leaving on February 1st, 2014" do
+                  setup do
+                    add_response "2014-02-01"
+                  end
+
+                  should "show the answer" do
+                    assert_current_node :estimated_tax_charge
+                    assert_state_variable :benefit_taxable_amount, "527.80"
+                    assert_state_variable :benefit_claimed_amount, "527.80"
+                    assert_state_variable :percentage_tax_charge, 100.0
+                    assert_state_variable :benefit_tax, "527"
+                  end
+                end
+              end
+            end
+          end
+        end
+      end
+    end
+  end
+
+
   context "additional calculation tests" do
     setup do
       add_response :income_work_out
@@ -468,7 +588,7 @@ class ChildBenefitTaxCalculatorTest < ActiveSupport::TestCase
       context "income at 60k" do
         setup do
           add_response "60000"
-          add_response :yes    
+          add_response :yes
         end
 
         should "ask about pension" do
@@ -478,8 +598,8 @@ class ChildBenefitTaxCalculatorTest < ActiveSupport::TestCase
         context "test 01" do
           setup do
             add_response "5000"    # Q3A Gross Pension
-            add_response "1600"    # Q4 net pension 
-            add_response "1000"    # Q5 trading losses 
+            add_response "1600"    # Q4 net pension
+            add_response "1000"    # Q5 trading losses
           end
 
           should "ask about charity donations" do
@@ -489,10 +609,10 @@ class ChildBenefitTaxCalculatorTest < ActiveSupport::TestCase
             assert_state_variable :net_pension_contributions, 1600
             assert_state_variable :trading_losses, 1000
             assert_state_variable :total_deductions, 8000
-            assert_state_variable :adjusted_net_income, 52000 
+            assert_state_variable :adjusted_net_income, 52000
           end
 
-          should "ask about children claiming for" do 
+          should "ask about children claiming for" do
             add_response "800"
             assert_current_node :how_many_children_claiming_for?
             assert_state_variable :adjusted_net_income, 51000
@@ -502,60 +622,60 @@ class ChildBenefitTaxCalculatorTest < ActiveSupport::TestCase
         context "test 02" do
           setup do
             add_response "5000"    # Q3A Gross Pension
-            add_response "0"    # Q4 net pension 
-            add_response "0"    # Q5 trading losses 
-            add_response "0"		# Q6 Gift Aided 
+            add_response "0"    # Q4 net pension
+            add_response "0"    # Q5 trading losses
+            add_response "0"		# Q6 Gift Aided
           end
 
-          should "ask about children claiming for" do 
+          should "ask about children claiming for" do
             assert_current_node :how_many_children_claiming_for?
             assert_state_variable :total_deductions, 5000
-            assert_state_variable :adjusted_net_income, 55000 
+            assert_state_variable :adjusted_net_income, 55000
           end
         end
 
         context "test 03" do
           setup do
             add_response "0"    # Q3A Gross Pension
-            add_response "1600"    # Q4 net pension 
-            add_response "0"    # Q5 trading losses 
-            add_response "0"		# Q6 Gift Aided 
+            add_response "1600"    # Q4 net pension
+            add_response "0"    # Q5 trading losses
+            add_response "0"		# Q6 Gift Aided
           end
 
-          should "ask about children claiming for" do 
+          should "ask about children claiming for" do
             assert_current_node :how_many_children_claiming_for?
             assert_state_variable :total_deductions, 2000
-            assert_state_variable :adjusted_net_income, 58000 
+            assert_state_variable :adjusted_net_income, 58000
           end
         end
-        
+
         context "test 04" do
           setup do
             add_response "0"    # Q3A Gross Pension
-            add_response "0"    # Q4 net pension 
-            add_response "3000"    # Q5 trading losses 
-            add_response "0"		# Q6 Gift Aided 
+            add_response "0"    # Q4 net pension
+            add_response "3000"    # Q5 trading losses
+            add_response "0"		# Q6 Gift Aided
           end
 
-          should "ask about children claiming for" do 
+          should "ask about children claiming for" do
             assert_current_node :how_many_children_claiming_for?
             assert_state_variable :total_deductions, 3000
-            assert_state_variable :adjusted_net_income, 57000 
+            assert_state_variable :adjusted_net_income, 57000
           end
         end
 
         context "test 05" do
           setup do
             add_response "0"    # Q3A Gross Pension
-            add_response "0"    # Q4 net pension 
-            add_response "0"    # Q5 trading losses 
-            add_response "1600"		# Q6 Gift Aided 
+            add_response "0"    # Q4 net pension
+            add_response "0"    # Q5 trading losses
+            add_response "1600"		# Q6 Gift Aided
           end
 
-          should "ask about children claiming for" do 
+          should "ask about children claiming for" do
             assert_current_node :how_many_children_claiming_for?
             assert_state_variable :total_deductions, 0
-            assert_state_variable :adjusted_net_income, 58000 
+            assert_state_variable :adjusted_net_income, 58000
           end
         end
       end
